add owner to water function

+++ backend/src/controllers/waters.js
@@ -1,12 +1,15 @@
 import { createWater, deleteWater, updateWater } from '../services/water.js';
+import { Water } from '../db/models/water.js';
 
 export const createWaterController = async (req, res) => {
   try {
-    const { date, amount, owner } = req.body;
+    const { date, amount } = req.body;
+    const { _id: owner } = req.user;
 
-    if (!date || !amount || !owner) {
+    if (!date || !amount) {
       return res.status(400).json({ message: 'Missing required fields' });
     }
+
     const newWater = await createWater(date, amount, owner);
     res.status(201).json(newWater);
   } catch (error) {
@@ -21,6 +24,18 @@ export const updateWaterController = async (req, res) => {
   try {
     const { waterId } = req.params;
     const { date, amount } = req.body;
+    const { _id: owner } = req.user;
+
+    if (!waterId) {
+      return res.status(400).json({ message: 'Missing waterId parameter' });
+    }
+
+    const waterRecord = await Water.findById(waterId);
+    if (!waterRecord || waterRecord.owner.toString() !== owner.toString()) {
+      return res
+        .status(403)
+        .json({ message: 'Access denied or record not found' });
+    }
 
     const updatedWater = await updateWater(waterId, { date, amount });
     res.status(200).json(updatedWater);
@@ -35,6 +50,18 @@ export const updateWaterController = async (req, res) => {
 export const deleteWaterController = async (req, res) => {
   try {
     const { waterId } = req.params;
+    const { _id: owner } = req.user;
+
+    if (!waterId) {
+      return res.status(400).json({ message: 'Missing waterId parameter' });
+    }
+
+    const waterRecord = await Water.findById(waterId);
+    if (!waterRecord || waterRecord.owner.toString() !== owner.toString()) {
+      return res
+        .status(403)
+        .json({ message: 'Access denied or record not found' });
+    }
 
     const deletedWater = await deleteWater(waterId);
     res.status(200).json(deletedWater);
